pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mebench"
version = "0.1.0"
description = "Paper-grade benchmark for comparing model extraction attacks fairly"
readme = "AGENTS.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Model Extraction Benchmark Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
mebench = "mebench.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mebench*"]

[tool.ruff]
target-version = "py38"
line-length = 100
select = [
    "E",
    "F",
    "W",
    "I",
]
ignore = []

[tool.ruff.format]
indent-style = "space"
quote-style = "double"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
